requirements.txt
Django==5.2
djangorestframework==3.16.0
django-htmx==1.23.0
ruff==0.11.8
pandas==2.2.3
pytest==8.3.5
pytest-django==4.11.1
